Fix funny performance bug
For some reason, I wrote the code to try all possible orders in which to
place pieces. That's silly, as the order of placement clearly doesn't
matter, only the position and orientation of each placement.

So the program ended up searching 7! = 5040 times the search space...
interesting that computers are now so fast that this didn't really matter
all that much...